Clean up some non-plot data
Add column indicating number of leaves in the scan.  This is for
lotus strigosus which had a few scans where all the leaflets were
lumped together.  Also extending the sequence column all the way
through all the scans.  Also adding dates for the DICA scans.

Add notes indicating leaves that I did that included the petiole
in the scan.

Change date format for my scans from April 20 and 21