Clean up species names and fix scans of leaf area
Leaf scans and leaf weights need to be merged by species, plant
leaf and whether or not the scan includes a petiole.  For some
species leaves were collected and scanned twice, once with the
petiole and then later without the petiole.  This and the
previous commits try to sort out the scans.